+++ apps/custom-config/wakeup-effect.js
@@ -4,7 +4,7 @@ var childProcess = require('child_process')
 var fs = require('fs')
 var promisify = require('util').promisify
 var ActivationConfig = require('@yoda/env')().activation
-var BaseConfig = require('./baseConfig')
+var BaseConfig = require('./base-config')
 var Caps = require('@yoda/caps/caps.node').Caps
 var property = require('@yoda/property')
 var safeParse = require('@yoda/util').json.safeParse
